+++ lib/src/core/better_player_with_controls.dart
@@ -160,9 +160,13 @@ class _BetterPlayerWithControlsState extends State<BetterPlayerWithControls> {
   ) {
     if (controlsConfiguration.showControls) {
       BetterPlayerTheme playerTheme = controlsConfiguration.playerTheme;
-      playerTheme ??= Platform.isAndroid != null
-          ? BetterPlayerTheme.material
-          : BetterPlayerTheme.cupertino;
+      if (playerTheme == null) {
+        if (Platform.isAndroid) {
+          playerTheme = BetterPlayerTheme.material;
+        } else {
+          playerTheme = BetterPlayerTheme.cupertino;
+        }
+      }
 
       if (controlsConfiguration.customControlsBuilder != null &&
           playerTheme == BetterPlayerTheme.custom) {